Decouple self-destruct from Slack proto
