+++ modules.gradle.kts
@@ -15,7 +15,6 @@ include("search:service")
 include("search:test-fixtures")
 include("self-destruct:adapter")
 include("self-destruct:domain")
-include("self-destruct:proto")
 include("self-destruct:service")
 include("slack:adapter")
 include("slack:domain")
