+++ self-destruct/adapter/build.gradle.kts
@@ -7,9 +7,7 @@ kotlin {
         val commonMain by getting {
             dependencies {
                 implementation(projects.common.network)
-                implementation(projects.slack.proto)
                 implementation(projects.selfDestruct.domain)
-                implementation(projects.selfDestruct.proto)
             }
         }
     }
