--- self-destruct/proto/build.gradle.kts
@@ -1,3 +1,0 @@
-plugins {
-    id("proto-plugin")
-}
