+++ self-destruct/service/build.gradle.kts
@@ -6,7 +6,6 @@ kotlin {
     sourceSets {
         val commonMain by getting {
             dependencies {
-                implementation(projects.slack.proto)
                 implementation(projects.selfDestruct.adapter)
                 implementation(projects.selfDestruct.domain)
             }
