+++ self-destruct/service/src/jsMain/kotlin/com/gchristov/thecodinglove/selfdestruct/service/SelfDestructService.kt
@@ -17,7 +17,6 @@ import com.gchristov.thecodinglove.selfdestruct.adapter.SelfDestructAdapterModul
 import com.gchristov.thecodinglove.selfdestruct.adapter.http.SelfDestructHttpHandler
 import com.gchristov.thecodinglove.selfdestruct.domain.SelfDestructDomainModule
 import com.gchristov.thecodinglove.selfdestruct.domain.model.Environment
-import com.gchristov.thecodinglove.slack.proto.SlackProtoModule
 
 suspend fun main() {
     // Ignore default Node arguments
@@ -44,7 +43,6 @@ private fun setupDi(environment: Environment): Either<Throwable, Unit> {
             CommonKotlinModule.module,
             CommonNetworkModule.module,
             CommonMonitoringModule(environment.apiUrl).module,
-            SlackProtoModule(environment.apiUrl).module,
             SelfDestructAdapterModule.module,
             SelfDestructDomainModule.module,
             SelfDestructServiceModule(environment).module,
